Creating extension point for kubernetesFetcher
