+++ plugins/kubernetes-backend/package.json
@@ -90,6 +90,8 @@
     "@backstage/backend-app-api": "workspace:^",
     "@backstage/backend-test-utils": "workspace:^",
     "@backstage/cli": "workspace:^",
+    "@backstage/plugin-permission-backend": "workspace:^",
+    "@backstage/plugin-permission-backend-module-allow-all-policy": "workspace:^",
     "@types/aws4": "^1.5.1",
     "msw": "^1.0.0",
     "supertest": "^6.1.3",
